requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.5
python-dotenv==1.1.0
Pillow==11.2.1
protobuf==4.25.3
google-generativeai==0.4.1